+++ brewtils/test/fixtures.py
@@ -359,7 +359,6 @@ def child_request_dict(ts_epoch):
         "created_at": ts_epoch,
         "updated_at": ts_epoch,
         "status_updated_at": ts_epoch,
-        "expiration_at": ts_epoch,
         "error_class": None,
         "metadata": {"child": "stuff"},
         "has_parent": True,
@@ -376,8 +375,6 @@ def child_request(child_request_dict, ts_dt):
     dict_copy["created_at"] = ts_dt
     dict_copy["updated_at"] = ts_dt
     dict_copy["status_updated_at"] = ts_dt
-    dict_copy["expiration_at"] = ts_dt
-
     return Request(**dict_copy)
 
 
@@ -405,7 +402,6 @@ def parent_request_dict(ts_epoch):
         "hidden": False,
         "updated_at": ts_epoch,
         "status_updated_at": ts_epoch,
-        "expiration_at": ts_epoch,
         "error_class": None,
         "metadata": {"parent": "stuff"},
         "has_parent": False,
@@ -422,7 +418,6 @@ def parent_request(parent_request_dict, ts_dt):
     dict_copy["created_at"] = ts_dt
     dict_copy["updated_at"] = ts_dt
     dict_copy["status_updated_at"] = ts_dt
-    dict_copy["expiration_at"] = ts_dt
     return Request(**dict_copy)
 
 
@@ -475,7 +470,6 @@ def request_dict(parent_request_dict, child_request_dict, ts_epoch):
         "created_at": ts_epoch,
         "updated_at": ts_epoch,
         "status_updated_at": ts_epoch,
-        "expiration_at": ts_epoch,
         "error_class": "ValueError",
         "metadata": {"request": "stuff"},
         "has_parent": True,
@@ -494,7 +488,6 @@ def bg_request(request_dict, parent_request, child_request, ts_dt):
     dict_copy["created_at"] = ts_dt
     dict_copy["updated_at"] = ts_dt
     dict_copy["status_updated_at"] = ts_dt
-    dict_copy["expiration_at"] = ts_dt
     return Request(**dict_copy)