+++ CHANGELOG.rst
@@ -16,6 +16,7 @@ TBD
 - Updated command signature validation error messages to include command name and parameter name
 - Fixed bug generating MD5 Hash from SystemClient uploaded Base64 Parameters
 - Fixed bug for Local System Client that doesn't load System object prior to executing command
+- Updated Local Request Handler to emulate all stages of Request lifecycle
 
 3.29.1
 ------