+++ CHANGELOG.rst
@@ -11,6 +11,8 @@ TBD
 - Added Entry Heartbeat event label
 - Updated StatusInfo heartbeat to append no more than one 'NOT_CONFIGURED' to history
 - Removed status, status info, and namespaces from Garden model since status is now derived from connections
+- Added expiration_at and root_command_type to Request model
+- Added Garden Name to System Model (#556)
 - Added root_command_type to Request model
 - Added updated_at/created_at to File and File Chunk model
 - Updated command signature validation error messages to include command name and parameter name